TEST lifecycle http manually

+++ src/test/rgw/bucket_notification/test_bn.py
@@ -410,7 +410,7 @@ kafka_server = 'localhost'
 
 class KafkaReceiver(object):
     """class for receiving and storing messages on a topic from the kafka broker"""
-    def __init__(self, topic, security_type):
+    def __init__(self, topic, security_type, kafka=None):
         from kafka import KafkaConsumer
         remaining_retries = 10
         port = 9092
@@ -1729,9 +1729,9 @@ def lifecycle(endpoint_type, conn, number_of_objects, topic_events, create_threa
     port = None
     if endpoint_type == 'http':
         # create random port for the http server
-        port = random.randint(10000, 20000)
+        port = 1824
         # start an http server in a separate thread
-        receiver = HTTPServerWithEvents((host, port))
+        # receiver = HTTPServerWithEvents((host, port))
         endpoint_address = 'http://'+host+':'+str(port)
         endpoint_args = 'push-endpoint='+endpoint_address+'&persistent=true'
     elif endpoint_type == 'amqp':
@@ -1798,25 +1798,27 @@ def lifecycle(endpoint_type, conn, number_of_objects, topic_events, create_threa
     no_keys = list(bucket.list())
     wait_for_queue_to_drain(topic_name, http_port=port)
     assert_equal(len(no_keys), 0)
-    event_keys = []
-    events = receiver.get_and_reset_events()
-    if not expected_abortion:
-        assert number_of_objects * 2 <= len(events)
-    for event in events:
-        assert_in(event['Records'][0]['eventName'], record_events)
-        event_keys.append(event['Records'][0]['s3']['object']['key'])
-    for key in keys:
-        key_found = False
-        for event_key in event_keys:
-            if event_key == key.name:
-                key_found = True
-                break
-        if not key_found:
-            err = 'no lifecycle event found for key: ' + str(key)
-            log.error(events)
-            assert False, err
+    # event_keys = []
+    # events = receiver.get_and_reset_events()
+    # if not expected_abortion:
+    #     assert number_of_objects * 2 <= len(events)
+    # for event in events:
+    #     assert_in(event['Records'][0]['eventName'], record_events)
+    #     event_keys.append(event['Records'][0]['s3']['object']['key'])
+    # for key in keys:
+    #     key_found = False
+    #     for event_key in event_keys:
+    #         if event_key == key.name:
+    #             key_found = True
+    #             break
+    #     if not key_found:
+    #         err = 'no lifecycle event found for key: ' + str(key)
+    #         log.error(events)
+    #         assert False, err
 
-    # cleanup
+    # cleanup, for non http
+    if endpoint_type == 'http':
+        return
     for key in keys:
         key.delete()
     if not expected_abortion:
